+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next version
 - Add option to install Metrics Server
 - Fix "Inspect node pools" macro when using managed node groups
+- Support tagging nodes
 - Remove macro `Run Kubectl command` (natively supported in DSS 10.0.6)
 - Support spot instances in node groups